+++ lib/ui/BrowsePanel.dart
@@ -30,8 +30,9 @@ import 'package:illinois/ui/academics/StudentCourses.dart';
 import 'package:illinois/ui/athletics/AthleticsContentPanel.dart';
 import 'package:illinois/ui/canvas/CanvasCoursesListPanel.dart';
 import 'package:illinois/ui/canvas/GiesCanvasCoursesListPanel.dart';
+import 'package:illinois/ui/directory/DirectoryAccountsList.dart';
 import 'package:illinois/ui/messages/MessagesHomePanel.dart';
-import 'package:illinois/ui/profile/ProfileDirectoryAccountsPanel.dart';
+import 'package:illinois/ui/directory/DirectoryAccountsPanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
 import 'package:illinois/ui/dining/DiningHomePanel.dart';
 import 'package:illinois/ui/gies/CheckListPanel.dart';
@@ -52,7 +53,6 @@ import 'package:illinois/ui/parking/ParkingEventsPanel.dart';
 import 'package:illinois/ui/polls/CreatePollPanel.dart';
 import 'package:illinois/ui/polls/CreateStadiumPollPanel.dart';
 import 'package:illinois/ui/polls/PollsHomePanel.dart';
-import 'package:illinois/ui/profile/ProfileDirectoryPage.dart';
 import 'package:illinois/ui/research/ResearchProjectsHomePanel.dart';
 import 'package:illinois/ui/safety/SafetyHomePanel.dart';
 import 'package:illinois/ui/surveys/PublicSurveysPanel.dart';
@@ -171,16 +171,6 @@ class _BrowseContentWidgetState extends State<BrowseContentWidget> implements No
     ]);
 
     _contentCodes = buildContentCodes();
-
-    // #4527 Auto-expand single item sections
-    if (_contentCodes != null) {
-      for (String code in _contentCodes!) {
-        List<String>? entryCodes = _BrowseSection.buildBrowseEntryCodes(sectionId: code);
-        if (entryCodes?.length == 1) {
-          _expandedCodes.add(code);
-        }
-      }
-    }
 
     super.initState();
   }
@@ -336,7 +326,7 @@ class _BrowseSection extends StatelessWidget {
 
   Widget _buildHeading(BuildContext context) {
     return Padding(padding: EdgeInsets.only(bottom: 4), child:
-      InkWell(onTap: _onTapExpand, child:
+      InkWell(onTap: () => _onTapHeading(context), child:
         Container(
           decoration: BoxDecoration(color: Styles().colors.white, border: Border.all(color: Styles().colors.surfaceAccent, width: 1),),
           padding: EdgeInsets.only(left: 16),
@@ -368,11 +358,7 @@ class _BrowseSection extends StatelessWidget {
                   Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 16), child:
                     SizedBox(width: 18, height: 18, child:
                       Center(child:
-                        _hasBrowseContent ? (
-                          expanded ?
-                            Styles().images.getImage('chevron-up', excludeFromSemantics: true) :
-                            Styles().images.getImage('chevron-down', excludeFromSemantics: true)
-                        ) : Container()
+                        _headingIcon
                       ),
                     )
                   ),
@@ -384,9 +370,27 @@ class _BrowseSection extends StatelessWidget {
     );
   }
 
+  Widget? get _headingIcon {
+    if (_hasBrowseContent) {
+      if (_singleBrowseCode != null) {
+        return Styles().images.getImage('chevron-right', excludeFromSemantics: true);
+      }
+      else if (expanded) {
+        return Styles().images.getImage('chevron-up', excludeFromSemantics: true);
+      }
+      else {
+        return Styles().images.getImage('chevron-down', excludeFromSemantics: true);
+      }
+    }
+    else {
+      return Container();
+    }
+  }
+
   Widget _buildEntries(BuildContext context) {
       List<Widget> entriesList = <Widget>[];
-      if (expanded && (_browseEntriesCodes != null)) {
+      int browseEntriesCount = expanded ? (_browseEntriesCodes?.length ?? 0) : 0;
+      if (1 < browseEntriesCount) {
         for (String code in _browseEntriesCodes!) {
           entriesList.add(_BrowseEntry(
             sectionId: sectionId,
@@ -412,13 +416,20 @@ class _BrowseSection extends StatelessWidget {
   static String description({required String sectionId}) =>
       AppTextUtils.appBrandString('panel.browse.section.$sectionId.description', '');
 
-  void _onTapExpand() {
-    if (_hasBrowseContent && (onExpand != null)) {
-      onExpand!();
+  void _onTapHeading(BuildContext context) {
+    if (_hasBrowseContent) {
+      String? singleBrowseCode = _singleBrowseCode;
+      if (singleBrowseCode != null) {
+        _BrowseEntry.process(context, sectionId, singleBrowseCode);
+      }
+      else {
+        onExpand?.call();
+      }
     }
   }
 
   bool get _hasBrowseContent => _browseEntriesCodes?.isNotEmpty ?? false;
+  String? get _singleBrowseCode => (_browseEntriesCodes?.length == 1) ? _browseEntriesCodes?.first : null;
 
   bool get _hasFavoriteContent {
     if (_browseEntriesCodes?.isNotEmpty ?? false) {
@@ -579,6 +590,10 @@ class _BrowseEntry extends StatelessWidget {
     Styles().images.getImage(_iconsMap['$sectionId.$entryId'] ?? 'chevron-right-bold', excludeFromSemantics: true);
 
   void _onTap(BuildContext context) {
+    process(context, sectionId, entryId);
+  }
+
+  static void process(BuildContext context, String sectionId, String entryId) {
     switch("$sectionId.$entryId") {
       case "academics.gies_checklist":        _onTapGiesChecklist(context); break;
       case "academics.new_student_checklist": _onTapNewStudentChecklist(context); break;
@@ -687,47 +702,47 @@ class _BrowseEntry extends StatelessWidget {
     }
   }
 
-  void _onTapGiesChecklist(BuildContext context) {
+  static void _onTapGiesChecklist(BuildContext context) {
     Analytics().logSelect(target: "Gies Checklist");
     CheckListPanel.present(context, contentKey: CheckList.giesOnboarding, analyticsFeature: AnalyticsFeature.AcademicsGiesChecklist);
   }
 
-  void _onTapNewStudentChecklist(BuildContext context) {
+  static void _onTapNewStudentChecklist(BuildContext context) {
     Analytics().logSelect(target: "New Student Checklist");
     CheckListPanel.present(context, contentKey: CheckList.uiucOnboarding, analyticsFeature: AnalyticsFeature.AcademicsChecklist);
   }
 
-  void _onTapSkillSelfEvaluation(BuildContext context) {
+  static void _onTapSkillSelfEvaluation(BuildContext context) {
     Analytics().logSelect(target: "Skills Self-Evaluation");
     AcademicsHomePanel.push(context, AcademicsContent.skills_self_evaluation);
   }
 
-  void _onTapEssentialSkillCoach(BuildContext context) {
+  static void _onTapEssentialSkillCoach(BuildContext context) {
     Analytics().logSelect(target: "Essential Skills Coach");
     AcademicsHomePanel.push(context, AcademicsContent.essential_skills_coach);
   }
 
-  void _onTapAcademicsToDo(BuildContext context) {
+  static void _onTapAcademicsToDo(BuildContext context) {
     Analytics().logSelect(target: "Academics To Do");
     AcademicsHomePanel.push(context, AcademicsContent.todo_list);
   }
 
-  void _onTapCanvasCourses(BuildContext context) {
+  static void _onTapCanvasCourses(BuildContext context) {
     Analytics().logSelect(target: "Canvas Courses");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCoursesListPanel()));
   }
 
-  void _onTapGiesCanvasCourses(BuildContext context) {
+  static void _onTapGiesCanvasCourses(BuildContext context) {
     Analytics().logSelect(target: "Gies Canvas Courses");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GiesCanvasCoursesListPanel()));
   }
 
-  void _onTapStudentCourses(BuildContext context) {
+  static void _onTapStudentCourses(BuildContext context) {
     Analytics().logSelect(target: "Student Courses");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => StudentCoursesListPanel()));
   }
 
-  void _onTapCampusReminders(BuildContext context) {
+  static void _onTapCampusReminders(BuildContext context) {
     Analytics().logSelect(target: "Campus Reminders");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GuideListPanel(
       contentList: Guide().remindersList,
@@ -737,30 +752,26 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  int get _videoTutorialsCount => uiuc.Content().videos?.length ?? 0;
-
-  bool get _canVideoTutorials => (_videoTutorialsCount > 0);
-
-  void _onTapVideoTutorials(BuildContext context) {
+  static void _onTapVideoTutorials(BuildContext context) {
     if (Connectivity().isOffline) {
       AppAlert.showOfflineMessage(context, Localization().getStringEx('panel.browse.label.offline.video_tutorial', 'Video Tutorial not available while offline.'));
     }
-    else if (_canVideoTutorials) {
+    else if (uiuc.Content().videos?.isNotEmpty == true) {
       List<Video>? videoTutorials = _getVideoTutorials();
       if (videoTutorials?.length == 1) {
         Video? videoTutorial = videoTutorials?.first;
         if (videoTutorial != null) {
-          Analytics().logSelect(target: "Video Tutorials", source: runtimeType.toString(), attributes: videoTutorial.analyticsAttributes);
+          Analytics().logSelect(target: "Video Tutorials", attributes: videoTutorial.analyticsAttributes);
           Navigator.push(context, CupertinoPageRoute( settings: RouteSettings(), builder: (context) => AppHelpVideoTutorialPanel(videoTutorial: videoTutorial)));
         }
       } else {
-        Analytics().logSelect(target: "Video Tutorials", source: runtimeType.toString());
+        Analytics().logSelect(target: "Video Tutorials",);
         Navigator.push(context, CupertinoPageRoute(settings: RouteSettings(), builder: (context) => AppHelpVideoTutorialListPanel(videoTutorials: videoTutorials)));
       }
     }
   }
 
-  List<Video>? _getVideoTutorials() {
+  static List<Video>? _getVideoTutorials() {
     Map<String, dynamic>? videoTutorials = uiuc.Content().videoTutorials;
     if (videoTutorials == null) {
       return null;
@@ -773,15 +784,13 @@ class _BrowseEntry extends StatelessWidget {
     return Video.listFromJson(jsonList: videos, contentStrings: strings);
   }
 
-  bool get _canFeedback => StringUtils.isNotEmpty(Config().feedbackUrl);
-
-  void _onTapFeedback(BuildContext context) {
+  static void _onTapFeedback(BuildContext context) {
     Analytics().logSelect(target: "Provide Feedback");
 
     if (Connectivity().isOffline) {
       AppAlert.showOfflineMessage(context, Localization().getStringEx('widget.home.app_help.feedback.label.offline', 'Providing a Feedback is not available while offline.'));
     }
-    else if (_canFeedback) {
+    else if (StringUtils.isNotEmpty(Config().feedbackUrl)) {
       String email = Uri.encodeComponent(Auth2().email ?? '');
       String name =  Uri.encodeComponent(Auth2().fullName ?? '');
       String phone = Uri.encodeComponent(Auth2().phone ?? '');
@@ -790,61 +799,59 @@ class _BrowseEntry extends StatelessWidget {
     }
   }
 
-  void _onTapReview(BuildContext context) {
+  static void _onTapReview(BuildContext context) {
     Analytics().logSelect(target: "Provide Review");
     InAppReview.instance.openStoreListing(appStoreId: Config().appStoreId);
   }
 
-  bool get _canFAQs => StringUtils.isNotEmpty(Config().faqsUrl);
-
-  void _onTapFAQs(BuildContext context) {
+  static void _onTapFAQs(BuildContext context) {
     Analytics().logSelect(target: "FAQs");
 
-    if (_canFAQs) {
+    if (StringUtils.isNotEmpty(Config().faqsUrl)) {
       _launchUrl(context, Config().faqsUrl);
     }
   }
 
-  void _onTapSportEvents(BuildContext context) {
+  static void _onTapSportEvents(BuildContext context) {
     Analytics().logSelect(target: "Events");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.events)));
   }
 
-  void _onTapSportNews(BuildContext context) {
+  static void _onTapSportNews(BuildContext context) {
     Analytics().logSelect(target: "News");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.news)));
   }
 
-  void _onTapSportTeams(BuildContext context) {
+  static void _onTapSportTeams(BuildContext context) {
     Analytics().logSelect(target: "Teams");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.teams)));
   }
 
-  void _onTapBuildingAccess(BuildContext context) {
+  static void _onTapBuildingAccess(BuildContext context) {
     Analytics().logSelect(target: 'Building Access');
     WalletICardHomeContentPanel.present(context, content: WalletICardContent.i_card);
   }
 
-  void _onTapTestLocations(BuildContext context) {
+  static void _onTapTestLocations(BuildContext context) {
     Analytics().logSelect(target: 'Locations');
     Navigator.push(context, CupertinoPageRoute(
       builder: (context) => HomeSaferTestLocationsPanel()
     ));
   }
 
-  void _onTapMyMcKinley(BuildContext context) {
+  static void _onTapMyMcKinley(BuildContext context) {
     Analytics().logSelect(target: 'MyMcKinley');
     _launchUrl(context, Config().saferMcKinleyUrl);
   }
 
-  void _onTapWellnessAnswerCenter(BuildContext context) {
+  static void _onTapWellnessAnswerCenter(BuildContext context) {
     Analytics().logSelect(target: 'Answer Center');
     Navigator.push(context, CupertinoPageRoute(
       builder: (context) => HomeSaferWellnessAnswerCenterPanel()
     ));
   }
 
-  void _onTapCampusHighlights(BuildContext context) {
+  static void _onTapCampusHighlights(BuildContext context) {
     Analytics().logSelect(target: 'Campus Highlights');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GuideListPanel(
       contentList: Guide().promotedList,
@@ -854,7 +861,7 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  void _onTapCampusSafetyResources(BuildContext context) {
+  static void _onTapCampusSafetyResources(BuildContext context) {
     Analytics().logSelect(target: 'Campus Safety Resources');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GuideListPanel(
       contentList: Guide().safetyResourcesList,
@@ -864,24 +871,21 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  bool get _canDueDateCatalog => StringUtils.isNotEmpty(Config().dateCatalogUrl);
-
-  void _onTapDueDateCatalog(BuildContext context) {
+  static void _onTapDueDateCatalog(BuildContext context) {
     Analytics().logSelect(target: "Due Date Catalog");
-
-    if (_canDueDateCatalog) {
+    if (StringUtils.isNotEmpty(Config().dateCatalogUrl)) {
       _launchUrl(context, Config().dateCatalogUrl);
     }
   }
 
-  void _onTapDiningsAll(BuildContext context) {
+  static void _onTapDiningsAll(BuildContext context) {
     Analytics().logSelect(target: "HomeDiningWidget: Residence Hall Dining");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => DiningHomePanel(
       analyticsFeature: AnalyticsFeature.DiningAll
     )));
   }
 
-  void _onTapDiningsOpen(BuildContext context) {
+  static void _onTapDiningsOpen(BuildContext context) {
     Analytics().logSelect(target: "HomeDiningWidget: Residence Hall Dining Open Now");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => DiningHomePanel(
       initialFilter: DiningFilter(type: DiningFilterType.work_time, selectedIndexes: {1}),
@@ -889,45 +893,45 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  void _onTapUserDirectory(BuildContext context) {
+  static void _onTapUserDirectory(BuildContext context) {
     Analytics().logSelect(target: "User Directory");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) { return ProfileDirectoryAccountsPanel(DirectoryAccounts.userDirectory); } ));
+    Navigator.push(context, CupertinoPageRoute(builder: (context) { return DirectoryAccountsPanel(DirectoryAccounts.directory); } ));
   }
 
-  void _onTapLaundry(BuildContext context) {
+  static void _onTapLaundry(BuildContext context) {
     Analytics().logSelect(target: "Laundry");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => LaundryHomePanel()));
   }
 
-  void _onTapMessages(BuildContext context) {
+  static void _onTapMessages(BuildContext context) {
     Analytics().logSelect(target: "Messages");
     MessagesHomePanel.present(context);
   }
 
-  void _onTapMTDStops(BuildContext context) {
+  static void _onTapMTDStops(BuildContext context) {
     Analytics().logSelect(target: "All Bus Stops");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => MTDStopsHomePanel(contentType: MTDStopsContentType.all,)));
   }
 
-  void _onTapCampusGuide(BuildContext context) {
+  static void _onTapCampusGuide(BuildContext context) {
     Analytics().logSelect(target: "Campus Guide");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CampusGuidePanel()));
   }
 
-  void _onTapNotifications(BuildContext context, {bool? unread}) {
+  static void _onTapNotifications(BuildContext context, {bool? unread}) {
     bool isUnread = (unread == true);
     Analytics().logSelect(target: isUnread ? "Unread Notifications" : "All Notifications");
     NotificationsHomePanel.present(context, content: isUnread ? NotificationsContent.unread : NotificationsContent.all);
   }
 
-  void _onTapEventFeed(BuildContext context) {
+  static void _onTapEventFeed(BuildContext context) {
     Analytics().logSelect(target: "Events Feed");
     Event2HomePanel.present(context,
       analyticsFeature: AnalyticsFeature.EventsAll,
     );
   }
 
-  void _onTapMyEvents(BuildContext context) {
+  static void _onTapMyEvents(BuildContext context) {
     Analytics().logSelect(target: "My Events");
     Event2HomePanel.present(context,
       types: LinkedHashSet<Event2TypeFilter>.from([Event2TypeFilter.favorite]),
@@ -935,127 +939,127 @@ class _BrowseEntry extends StatelessWidget {
     );
   }
 
-  void _onTapTwitter(BuildContext context) {
+  static void _onTapTwitter(BuildContext context) {
     Analytics().logSelect(target: "Twitter");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return TwitterPanel(); } ));
   }
 
-  void _onTapDailyIllini(BuildContext context) {
+  static void _onTapDailyIllini(BuildContext context) {
     Analytics().logSelect(target: "Daily Illini");
     _launchUrl(context, Config().dailyIlliniHomepageUrl);
   }
 
-  void _onTapRadioStation(BuildContext context, RadioStation radioStation) {
+  static void _onTapRadioStation(BuildContext context, RadioStation radioStation) {
     Analytics().logSelect(target: "Radio Station: ${RadioPopupWidget.stationTitle(radioStation)}");
     RadioPopupWidget.show(context, radioStation);
   }
 
-  void _onTapAllGroups(BuildContext context) {
+  static void _onTapAllGroups(BuildContext context) {
     Analytics().logSelect(target: "All Groups");
     Navigator.push(context, CupertinoPageRoute(settings: RouteSettings(name: GroupsHomePanel.routeName), builder: (context) => GroupsHomePanel(contentType: GroupsContentType.all,)));
   }
 
-  void _onTapMyGroups(BuildContext context) {
+  static void _onTapMyGroups(BuildContext context) {
     Analytics().logSelect(target: "My Groups");
     Navigator.push(context, CupertinoPageRoute(settings: RouteSettings(name: GroupsHomePanel.routeName), builder: (context) => GroupsHomePanel(contentType: GroupsContentType.my)));
   }
 
-  void _onTapOpenResearchProjects(BuildContext context) {
+  static void _onTapOpenResearchProjects(BuildContext context) {
     Analytics().logSelect(target: "Open Research Projects");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => ResearchProjectsHomePanel(contentType: ResearchProjectsContentType.open,)));
   }
 
-  void _onTapMyResearchProjects(BuildContext context) {
+  static void _onTapMyResearchProjects(BuildContext context) {
     Analytics().logSelect(target: "My Research Projects");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => ResearchProjectsHomePanel(contentType: ResearchProjectsContentType.my)));
   }
 
-  void _onTapMyGameDay(BuildContext context) {
+  static void _onTapMyGameDay(BuildContext context) {
     Analytics().logSelect(target: "It's Game Day");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.game_day)));
   }
 
-  void _onTapMyDinings(BuildContext context) {
+  static void _onTapMyDinings(BuildContext context) {
     Analytics().logSelect(target: "My Dinings");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [Dining.favoriteKeyName]); } ));
   }
 
-  void _onTapMyAthletics(BuildContext context) {
+  static void _onTapMyAthletics(BuildContext context) {
     Analytics().logSelect(target: "My Big 10 Events");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.my_events)));
   }
 
-  void _onTapMyNews(BuildContext context) {
+  static void _onTapMyNews(BuildContext context) {
     Analytics().logSelect(target: "My News");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsContentPanel(content: AthleticsContent.my_news)));
   }
 
-  void _onTapMyLaundry(BuildContext context) {
+  static void _onTapMyLaundry(BuildContext context) {
     Analytics().logSelect(target: "My Laundry");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [LaundryRoom.favoriteKeyName]); } ));
   }
 
-  void _onTapMyMTDStops(BuildContext context) {
+  static void _onTapMyMTDStops(BuildContext context) {
     Analytics().logSelect(target: "My Bus Stops");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => MTDStopsHomePanel(contentType: MTDStopsContentType.my,)));
   }
 
-  void _onTapMyLocations(BuildContext context) {
+  static void _onTapMyLocations(BuildContext context) {
     Analytics().logSelect(target: "My Locations");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [ExplorePOI.favoriteKeyName]); } ));
   }
 
-  void _onTapMyCampusGuide(BuildContext context) {
+  static void _onTapMyCampusGuide(BuildContext context) {
     Analytics().logSelect(target: "My Campus Guide");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [GuideFavorite.favoriteKeyName]); } ));
   }
 
-  void _onTapWellnessResources(BuildContext context) {
+  static void _onTapWellnessResources(BuildContext context) {
     Analytics().logSelect(target: "Wellness Resources");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return WellnessHomePanel(content: WellnessContent.resources,); } ));
   }
 
-  void _onTapWellnessMentalHealth(BuildContext context) {
+  static void _onTapWellnessMentalHealth(BuildContext context) {
     Analytics().logSelect(target: "Wellness Resources");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return WellnessHomePanel(content: WellnessContent.mentalHealth,); } ));
   }
 
-  void _onTapWellnessAppointments(BuildContext context) {
+  static void _onTapWellnessAppointments(BuildContext context) {
     Analytics().logSelect(target: "MyMcKinley Appointments");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return WellnessHomePanel(content: WellnessContent.appointments); } ));
   }
 
-  void _onTapAppointments(BuildContext context, { AnalyticsFeature? analyticsFeature }) {
+  static void _onTapAppointments(BuildContext context, { AnalyticsFeature? analyticsFeature }) {
     Analytics().logSelect(target: "Appointments");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => AppointmentsListPanel(analyticsFeature: analyticsFeature,)));
   }
 
-  void _onTapAcademicsMyIllini(BuildContext context) {
+  static void _onTapAcademicsMyIllini(BuildContext context) {
     Analytics().logSelect(target: "myIllini");
     _launchUrl(context, Config().myIlliniUrl);
   }
 
-  void _onTapCreatePoll(BuildContext context) {
+  static void _onTapCreatePoll(BuildContext context) {
     Analytics().logSelect(target: "Create Poll");
     CreatePollPanel.present(context);
   }
 
-  void _onTapViewPolls(BuildContext context) {
+  static void _onTapViewPolls(BuildContext context) {
     Analytics().logSelect(target: "View Polls");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => PollsHomePanel()));
   }
 
-  void _onTapRecentItems(BuildContext context) {
+  static void _onTapRecentItems(BuildContext context) {
     Analytics().logSelect(target: "Recent Items");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => HomeRecentItemsPanel()));
   }
 
-  void _onTapSafewalkRequest(BuildContext context) {
+  static void _onTapSafewalkRequest(BuildContext context) {
     Analytics().logSelect(target: "Request a SafeWalk");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => SafetyHomePanel()));
   }
 
-  void _onTapSafeRides(BuildContext context) {
+  static void _onTapSafeRides(BuildContext context) {
     Analytics().logSelect(target: "SafeRides (MTD)");
     Map<String, dynamic>? safeRidesGuideEntry = Guide().entryById(Config().safeRidesGuideId);
     if (safeRidesGuideEntry != null) {
@@ -1063,7 +1067,7 @@ class _BrowseEntry extends StatelessWidget {
     }
   }
 
-  void _onTapSafetyResources(BuildContext context) {
+  static void _onTapSafetyResources(BuildContext context) {
     Analytics().logSelect(target: "Safety Resources");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => GuideListPanel(
       contentList: Guide().safetyResourcesList,
@@ -1073,17 +1077,17 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  void _onTapPublicSurveys(BuildContext context) {
+  static void _onTapPublicSurveys(BuildContext context) {
     Analytics().logSelect(target: "Public Surveys");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => PublicSurveysPanel()));
   }
 
-  void _onTapParking(BuildContext context) {
+  static void _onTapParking(BuildContext context) {
     Analytics().logSelect(target: "Parking");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => ParkingEventsPanel()));
   }
 
-  void _onTapStateFarmWayfinding(BuildContext context) {
+  static void _onTapStateFarmWayfinding(BuildContext context) {
     Analytics().logSelect(target: "State Farm Wayfinding");
     /* TBD Map2 NativeCommunicator().launchMap(target: {
       'latitude': Config().stateFarmWayfinding['latitude'],
@@ -1092,77 +1096,77 @@ class _BrowseEntry extends StatelessWidget {
     }); */
   }
 
-  void _onTapCreateStadiumPoll(BuildContext context) {
+  static void _onTapCreateStadiumPoll(BuildContext context) {
     Analytics().logSelect(target: "Create Stadium Poll");
     CreateStadiumPollPanel.present(context);
   }
 
-  void _onTapIlliniCash(BuildContext context) {
+  static void _onTapIlliniCash(BuildContext context) {
     Analytics().logSelect(target: "Illini Cash");
     WalletHomePanel.present(context, contentType: WalletContentType.illiniCash);
   }
 
-  void _onTapAddIlliniCash(BuildContext context) {
+  static void _onTapAddIlliniCash(BuildContext context) {
     Analytics().logSelect(target: "Add Illini Cash");
     WalletHomePanel.present(context, contentType: WalletContentType.addIlliniCash);
   }
 
-  void _onTapMealPlan(BuildContext context) {
+  static void _onTapMealPlan(BuildContext context) {
     Analytics().logSelect(target: "Meal Plan");
     WalletHomePanel.present(context, contentType: WalletContentType.mealPlan);
   }
 
-  void _onTapBusPass(BuildContext context) {
+  static void _onTapBusPass(BuildContext context) {
     Analytics().logSelect(target: "Bus Pass");
     WalletHomePanel.present(context, contentType: WalletContentType.busPass);
   }
 
-  void _onTapIlliniId(BuildContext context) {
+  static void _onTapIlliniId(BuildContext context) {
     Analytics().logSelect(target: "Illini ID");
     WalletHomePanel.present(context, contentType: WalletContentType.illiniId);
   }
 
-  void _onTapLibraryCard(BuildContext context) {
+  static void _onTapLibraryCard(BuildContext context) {
     Analytics().logSelect(target: "Library Card");
     _notImplemented(context);
   }
 
-  void _onTapWellnessRings(BuildContext context) {
+  static void _onTapWellnessRings(BuildContext context) {
     Analytics().logSelect(target: "Wellness Daily Rings");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.rings,)));
   }
 
-  void _onTapWellnessToDo(BuildContext context) {
+  static void _onTapWellnessToDo(BuildContext context) {
     Analytics().logSelect(target: "Wellness To Do");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.todo,)));
   }
 
-  void _onTapWellnessTips(BuildContext context) {
+  static void _onTapWellnessTips(BuildContext context) {
     Analytics().logSelect(target: "Wellness Daily Tips");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.dailyTips,)));
   }
 
-  void _onTapWellnessHealthScreener(BuildContext context) {
+  static void _onTapWellnessHealthScreener(BuildContext context) {
     Analytics().logSelect(target: "Illinois Health Screener");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.healthScreener,)));
   }
 
-  void _onTapWellnessSuccessTeam(BuildContext context) {
+  static void _onTapWellnessSuccessTeam(BuildContext context) {
     Analytics().logSelect(target: "My Success Team");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.successTeam,)));
   }
 
-  void _onTapWellnessPodcast(BuildContext context) {
+  static void _onTapWellnessPodcast(BuildContext context) {
     Analytics().logSelect(target: "Healthy Illini Podcast");
     _launchUrl(context, Wellness().getResourceUrl(resourceId: 'podcast'));
   }
 
-  void _onTapWellnessStruggling(BuildContext context) {
+  static void _onTapWellnessStruggling(BuildContext context) {
     Analytics().logSelect(target: "I'm Struggling");
     _launchUrl(context, Wellness().getResourceUrl(resourceId: 'where_to_start'));
   }
 
-  void _notImplemented(BuildContext context) {
+  static void _notImplemented(BuildContext context) {
     AppAlert.showDialogResult(context, "Not implemented yet.");
   }
 
@@ -1182,7 +1186,6 @@ class _BrowseEntry extends StatelessWidget {
       }
     }
   }
-
 }
 
 ///////////////////////////